renamed, reorged, updated examples and tests
