+++ examples/arrange.rs
@@ -10,7 +10,7 @@ use timely::dataflow::operators::*;
 use rand::{Rng, SeedableRng, StdRng};
 
 use differential_dataflow::AsCollection;
-use differential_dataflow::operators::arrange_alt::ArrangeByKey;
+use differential_dataflow::operators::arrange::ArrangeByKey;
 use differential_dataflow::trace::{Cursor, Trace};
 
 fn main() {
